requirements.txt
Django==2.2.4
djangorestframework==3.9.4
pygments
django-rest-swagger
jsonfield
gunicorn
coreapi
drf-extensions
textblob
nltk
pandas
sklearn
scikit-learn
whitenoise
dj_static
lxml==4.2.2
scipy==1.1.0
matplotlib==2.2.2
django-cors-headers
psycopg2==2.7.5
psycopg2-binary==2.7.5
statsmodels
dj-database-url